requirements.txt
yfinance
faiss-cpu
openai
beautifulsoup4
sqlalchemy
pandas
numpy
pymysql
python-dotenv
lxml

